% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htestimate.R
\name{createProbMatrix}
\alias{createProbMatrix}
\title{Probabilities and joint probabilities of assignment.}
\usage{
createProbMatrix(raw_assignments, byrow = F)
}
\arguments{
\item{byrow}{Change the assignment matrix to use rows rather than columns for each randomization.}

\item{assignments}{Each column is a replication and each element is the unit's assignment.}
}
\value{
n*k by n*k matrix of probabilities and joint probabilities of assignment to each arm.
}
\description{
Return a matrix of probabilities and joint probabilities of assignment for n
observations and k randomization replications.
}
\examples{
TBD
}

