% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htEstimate.R
\name{generateAssignmentProbs}
\alias{generateAssignmentProbs}
\title{TBD}
\usage{
generateAssignmentProbs(row, col, assignments)
}
\arguments{
\item{row}{The assignment level to be used in the row check.}

\item{col}{The assignment level to be used in the column check.}

\item{assignments}{These have already been mapped to 1..k}
}
\value{
nxn matrix with joint assignment probabilities.
}
\description{
Used by createProbMatrix.
}

