% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htestimate.R
\name{getRawMatrixEntries}
\alias{getRawMatrixEntries}
\title{Return probability matrix locations for the sub-matrix for a certain observation pair.}
\usage{
getRawMatrixEntries(row, col, n, direct = F)
}
\arguments{
\item{row}{TBD}

\item{col}{TBD}

\item{n}{Total number of observations.}

\item{direct}{Not yet implemeneted: try to return the ranges for direct usage, rather than a list.}
}
\value{
TBD
}
\description{
TBD
}

