% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htEstimate.R
\name{matDiag}
\alias{matDiag}
\title{Return the diagonal of the matrix.}
\usage{
matDiag(mat)
}
\arguments{
\item{mat}{Input matrix, must be square.}
}
\value{
diagonal elements of the matrix as a vector.
}
\description{
TBD
}
\examples{
TBD
}

