% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htestimate.R
\name{matDiag}
\alias{matDiag}
\title{Return the diagonal of the matrix.}
\usage{
matDiag(mat)
}
\arguments{
\item{mat}{Input matrix, must be square.}
}
\value{
diagonal elements of the matrix as a vector.
}
\description{
Returns the diagonal elements of a matrix as a vector.
}
\examples{
myMatrix = matrix(1:9, ncol=3)
# This returns c(1, 5, 9)
diagonal = matDiag(myMatrix)
}

